Add secondsToDiffTime and picosecondsToDiffTime.
Rationale:
As has come up on haskell-cafe
(http://comments.gmane.org/gmane.comp.lang.haskell.cafe/15653), it
takes a while to figure out how to make DiffTime values.

secondsToDiffTime is not that important since it is just another name
for fromInteger, but I suspect that it would be used a lot.  Using
fromRational to create a DiffTime from a number of picoseconds is a
bit of a hassle, so having a picosecondsToDiffTime would be useful.

darcs-hash:20061115222145-6cdb2-4c0badc67fc5a5c1880c111902ee3e28ad793719